backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==0.8.11
PyMuPDF==1.23.8
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3
google-generativeai==0.3.2
fpdf2==2.7.6
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.1
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3